refactor: S2 split candidates.py + copilot.py + deploy conditionnel + tests services
- candidates.py (2462L) -> 7 fichiers dans candidates/ package
- copilot.py (932L) -> copilot.py + copilot_prompts.py + copilot_parser.py
- tests services: pdf_report, profile_compute, copilot_parser (49 tests)
- deploy.yml conditionnel au succes des tests
- fix deadlock DB dans test_cv_scoring (TestSyncSession)
